researchflow-production-main/agents/requirements.txt
# ResearchFlow Agent Dependencies
# Install: pip install -r agents/requirements.txt

# Core LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Composio Integration
composio-core>=0.3.0
composio-langchain>=0.3.0
composio-openai>=0.3.0

# Async Support
aiohttp>=3.9.2  # Security: CVE-2024-23334 fix
asyncio-throttle>=1.0.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.27.0  # Security: Updated for security fixes (async support for AI helper)
requests>=2.31.0

# YAML/JSON Processing
pyyaml>=6.0.0
orjson>=3.9.0

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-httpx>=0.30.0  # Mock HTTP requests in tests
pytest-cov>=4.1.0  # Code coverage

# Secret Management Backends (optional)
hvac>=1.2.1  # HashiCorp Vault client
boto3>=1.34.0  # AWS Secrets Manager (optional)
azure-keyvault-secrets>=4.7.0  # Azure Key Vault (optional)
azure-identity>=1.15.0  # Azure authentication (optional)

# Database Connectivity (for health checks)
psycopg2-binary>=2.9.9  # PostgreSQL
redis>=5.0.0  # Redis client

# Production Enhancements
prometheus_client>=0.20.0  # Metrics export
psutil>=5.9.0  # System resource monitoring
aiofiles>=23.2.0  # Async file operations
PyYAML>=6.0.1  # Configuration files

# Error Tracking & APM (optional)
sentry-sdk>=1.40.0  # Error tracking and performance monitoring
